src/test/java/com/ImageGallery/controller/ImageGalleryControllerTest.java
package com.ImageGallery.controller;

import static org.mockito.Mockito.*;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

import com.ImageGallery.service.ImageGalleryService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

@ExtendWith({SpringExtension.class, MockitoExtension.class})
@WebMvcTest(ImageGalleryController.class)
public class ImageGalleryControllerTest {

    private MockMvc mockMvc;

    @Mock
    private ImageGalleryService imageGalleryService;

    @InjectMocks
    private ImageGalleryController imageGalleryController;

    @BeforeEach
    public void setUp() {
        mockMvc = MockMvcBuilders.standaloneSetup(imageGalleryController).build();
    }

    @Test
    public void testDeleteImageGallery() throws Exception {
        int imageId = 1;
        String expectedResponse = "Image deleted successfully";

        when(imageGalleryService.deleteImageGallery(imageId)).thenReturn(expectedResponse);

        mockMvc.perform(delete("/api/images/{id}", imageId))
                .andExpect(status().isOk())
                .andExpect(content().string(expectedResponse));

        verify(imageGalleryService).deleteImageGallery(imageId);
    }
}